changing location for update issues on taskboard and add common steps for issues linking with milesotne
