+++ tests/acceptance/steps/common-steps/set-milestone-issues-steps.js
@@ -1,9 +1,5 @@
-import steps from './steps';
-import { triggerEvent } from '@ember/test-helpers';
+import steps from '../steps';
 import Collection from 'ember-cli-mirage/orm/collection';
-
-// step definitions that are shared between features should be moved to the
-// tests/acceptance/steps/steps.js file
 
 export const given = function () {
     return [
@@ -47,23 +43,6 @@ export const given = function () {
                     return issueCollection;
                 }
             }
-        }
-    ];
-}
-
-export const when = function () {
-    return [
-        {
-            "User change status of issue $issueId to $status": () => async function (issueId, status) {
-                let item = document.querySelector(`div.item[data-field-issue-id="${issueId}"]`);
-                let issue = server.schema.issues.find(issueId);
-                issue.update({ 'status': status });
-
-                await triggerEvent(item, "pointerdown", {
-                    button: 0
-                });
-
-            },
         }
     ];
 }
